Remove services.ts file and integrate its functionality into index.ts
- Deleted the services.ts file, consolidating its content into index.ts for better organization.
- Retained all existing functions related to wallet creation, transaction handling, and token management.
- This change streamlines the codebase by reducing the number of files and improving accessibility to service functions.
